Update FV3 to GEFS v12 branch
In order to keep as much consistency as possible with already produced
retrospectives, GEFS v12 requires FV3 to branch off of the main line to
add certain changes but not others. Port updates to the FV3 master will
be cherry-picked to the GEFS v12 branch, but other commits will not.

Change-Id: If8786719cc1f0f0234bd6d26d4def75c6696b027
Refs: #69950