Update baseline and regtest logs for Cray [except 1]
Updated the baseline and regtest logs for all tests but one on Cray. The
cpld_1way@block_fv3@32bit@C384_ww3@glo30m test hangs and fails when the
wallclock is exhausted. This issue extends beyond this commit; previous
commits that previously passed that test now fail, and the same test also
fails in the UFS NEMS app. The missing baseline will be added when the issue
is resolved.

Change-Id: I1044d058fd8133e207a239901f0d3fab9dd56c99